[HMC/clean] Move PreviewSteps to its own header
